backend/requirements.txt
fastapi==0.128.8
uvicorn==0.39.0
pytest==8.3.3
httpx==0.27.2
pydantic==2.13.0
pydantic_core==2.46.0
starlette==0.49.3
typing_extensions==4.15.0
anyio==4.12.1
python-dotenv==1.0.1
pytest-asyncio==1.2.0
google-generativeai==0.8.3
openai==1.54.3
anthropic==0.39.0
